Reduce implementation code
